+++ src/Carpark.ts
@@ -1,0 +1,3 @@
+class Carpark {
+
+}
