+++ src/Tenant.ts
@@ -1,0 +1,11 @@
+class Tenant {
+    public TenantID;
+    public Forename;
+    public Surname;
+
+    constructor(TenantID: number, Forename: String, Surname: String) {
+        this.TenantID = TenantID;
+        this.Forename = Forename;
+        this.Surname = Surname;
+    }
+}
