+++ src/Vehicle.ts
@@ -1,0 +1,11 @@
+class Vehicle {
+    public VehicleID;
+    public Numberplate;
+    public TenantID;
+
+    constructor(VehicleID: number, Numberplate: String, TenantID: number) {
+        this.VehicleID = VehicleID;
+        this.Numberplate = Numberplate;
+        this.TenantID = TenantID;
+    }
+}
