+++ tsconfig.json
@@ -1,10 +1,19 @@
 {
-    "compilerOptions": {
-        "module": "commonjs",
-        "moduleResolution": "node",
-        "esModuleInterop": true,
-        "target": "es2017"17",
-        "watch": true
-    }
+	"compilerOptions": {
+		"module": "CommonJS",
+		"target": "es6",
+		// "esModuleInterop": true,
+		"allowSyntheticDefaultImports": true,
+		"watch": true,
+		"strict": true,
+		"alwaysStrict": true,
+		"rootDir": "src",
+		"outDir": "./dist"
+	},
+    "include": [
+        "**/*.ts"
+    ], 
+    "exclude": [
+        "node_modules"
+    ]
 }
-// https://www.digitalocean.com/community/tutorials/setting-up-a-node-project-with-typescript
